pyproject.toml
[project]
name = "chandra-client"
version = "0.1.0"
description = "Lightweight Python client for Chandra OCR with vLLM support. Fork of datalab-to/chandra."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "sljeff", email = "sljeff@users.noreply.github.com"}
]
keywords = ["ocr", "pdf", "markdown", "layout", "vllm", "client"]
dependencies = [
    "beautifulsoup4>=4.14.2",
    "click>=8.0.0",
    "filetype>=1.2.0",
    "markdownify==1.1.0",
    "openai>=2.2.0",
    "pillow>=10.2.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pypdfium2>=4.30.0",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["setuptools>=61"]  # or "setuptools>=61", "flit-core", etc.
build-backend = "setuptools.build_meta"



[tool.setuptools.packages.find]
include = ["chandra*"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
]
